feat: datazoom update callback supports new trigger tag param
